move tests, change logging
