+++ src/LoginServer/LoginServer.test.ts
@@ -5,11 +5,12 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-import { LoginServer } from "../../../LoginServer/LoginServer";
-import { DatabaseManager } from "../../../databaseManager";
+import { LoginServer } from "./LoginServer";
+import { DatabaseManager } from "../databaseManager";
 const database = new DatabaseManager();
 
 const loginServer = new LoginServer(database);
+loginServer.logger.level("warn");
 
 test("retuns a valid customer id - d316cd2dd6bf870893dfbaaf17f965884e", () => {
   const { customerId, userId } = loginServer._npsGetCustomerIdByContextId(
